pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "memsearch"
version = "0.3.1"
description = "Semantic memory search for markdown knowledge bases"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "pymilvus>=2.5.0,!=2.6.10",
    "milvus-lite>=2.5.0; sys_platform != 'win32'",
    "click>=8.1",
    "watchdog>=4.0",
    "setuptools<75",
    "tomli_w>=1.0",
    "tomli>=2.0; python_version < '3.11'",
    "openai>=1.0",
    "pydantic-ai-slim[google,openai]>=1.0",
]

[project.optional-dependencies]
google = ["google-genai>=1.0"]
voyage = ["voyageai>=0.3"]
jina = ["httpx>=0.27"]
mistral = ["mistralai>=1.0"]
ollama = [
    "ollama>=0.4",
]
local = [
    "einops>=0.8.2",
    "sentence-transformers>=3.0",
]
anthropic = ["anthropic>=0.40"]
onnx = [
    "onnxruntime>=1.17,<1.24; python_version == '3.10'",
    "onnxruntime>=1.17; python_version >= '3.11'",
    "tokenizers>=0.15",
    "huggingface-hub>=0.20",
]
all = ["memsearch[google,voyage,jina,mistral,ollama,local,anthropic,onnx]"]

[project.scripts]
memsearch = "memsearch.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/memsearch"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24", "ruff>=0.9", "pytest-cov>=6.0", "pre-commit>=4.0"]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.1",
    "mkdocs-terminal>=4.8.0",
]

[tool.ruff]
target-version = "py310"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "W",  # pycodestyle
    "F",       # pyflakes
    "I",       # isort
    "UP",      # pyupgrade
    "B",       # flake8-bugbear
    "C4",      # flake8-comprehensions
    "SIM",     # flake8-simplify
    "PIE",     # flake8-pie
    "PGH",     # pygrep-hooks
    "RUF",     # ruff-specific
    "T20",     # flake8-print
    "PERF",    # perflint
]
ignore = [
    "E501",    # line length handled by formatter
]

[tool.ruff.lint.isort]
known-first-party = ["memsearch"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["T20"]      # allow print in tests
"examples/*" = ["T20"]   # allow print in examples

[tool.coverage.run]
source = ["memsearch"]

[tool.coverage.report]
show_missing = true
skip_empty = true
